몬스터 애니메이션 구현
1. 상태에 따른 애니메이션 변경 구현
2. 탐지 component 구현

+++ Assets/03. Vecolius/1. Script/InterfaceGroup.cs
@@ -24,7 +24,7 @@ namespace Veco
         /// Character Hit IMethod
         /// </summary>
         /// <param name="damage"> Get Damage </param>
-        void Hit(float damage);
+        void Hit(int damage);
     }
     public interface IAttackable
     {
